Create tenderly.md
Tenderly is a product, and although this is more of a community-resource-driven initiative, we do have a very generous free plan with all features available from the start which proved to be very useful for a bunch of people battling with local setups in order to test anything properly with actual on-chain data. This should  be more than enough for anyone to do whatever they want short of supporting a live product with a significant user base. Hopefully you and others will find this set of tools useful and valid to be added to the list, as our primary goal is to further blockchain adoption as a whole.